refactor to status test
